work on bug fixing

+++ lib/presentation/features/mwp/data/SaveMWPModel.dart
@@ -18,7 +18,7 @@ class SaveMWPModel {
   String createdBy;
   String actionedBy;
 
-  int dspConversionVol;
+  double dspConversionVol;
   int contractorVisit;
   int technocratVisit;
   int techVanDemo;
